requirements.txt
accelerate==0.15.0
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bitsandbytes==0.36.0.post2
certifi==2022.12.7
charset-normalizer==2.1.1
comm==0.1.2
datasets==2.8.0
debugpy==1.5.1
decorator==5.1.1
diffusers==0.11.1
dill==0.3.6
entrypoints==0.4
executing==0.8.3
filelock==3.9.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.19.2
ipython==8.7.0
ipywidgets==8.0.4
jedi==0.18.1
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyterlab-widgets==3.0.5
matplotlib-inline==0.1.6
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
platformdirs==2.5.2
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
pyzmq==23.2.0
regex==2022.10.31
requests==2.28.1
responses==0.18.0
setuptools==65.5.0
six==1.16.0
stack-data==0.2.0
tokenizers==0.13.2
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
wheel==0.37.1
widgetsnbextension==4.0.5
xxhash==3.2.0
yarl==1.8.2
zipp==3.11.0
